requirements.txt
pyTelegramBotAPI
logging
python-dotenv
jinja2